Add Defign typography system
The `DefignFonts` class defines constants for font size and line height for every different variant of text available in Defign. Not only does it provide instances of all of the variants, but it also provides utility functions for applying the variant and it's relevant font size to your component's properties.

A DSL for this is also provided, which is recommended to be used at the tail-end of your `properties` block, such as shown below:
```kt
val text = TextComponent(SimpleTextHolder("Hello, World!")).configure {
    properties {
        fill = Color.WHITE.asProperty
    }.useFontAttention1()
}.attachTo(frame)
```